Move Headers to /components/Header
